requirements.txt
appnope
asttokens
autopep8
backcall
black
certifi
charset-normalizer
click
comm
contourpy
cycler
debugpy
decorator
et-xmlfile
executing
flake8
fonttools
idna
ipykernel
ipython
jedi
joblib
jupyter_client
jupyter_core
kiwisolver
lightgbm
matplotlib
matplotlib-inline
mccabe
mypy-extensions
nest-asyncio
numpy
openpyxl
packaging
pandas
parso
pathspec
pexpect
pickleshare
Pillow
platformdirs
prompt-toolkit
psutil
ptyprocess
pure-eval
pycodestyle
pyflakes
Pygments
pyparsing
python-dateutil
pytz
pyzmq
requests
scikit-learn
scipy
seaborn
six
#sklearn==0.0.post5
stack-data
threadpoolctl
tomli
tornado
traitlets
tubular
tzdata
urllib3
wcwidth
xgboost
fastapi
uvicorn
pytest
requests
azure-cli==2.43.0
azure-identity==1.12.0
azure-storage-blob==12.12.0
pyyaml==6.0